requirements/local.txt
# Local development dependencies go here
-r _base.txt

django-debug-toolbar==0.9.4
nose==1.3.6
django-nose==1.4
factory-boy==2.1.2
